Removing postgres_session and moving it to history, start to implement saving blocks
