+++ history/Cargo.toml
@@ -18,3 +18,11 @@ tokio = "1.*"
 
 solana-lite-rpc-core = {workspace = true}
 solana-rpc-client-api = {workspace = true}
+native-tls = { workspace = true }
+postgres-native-tls = { workspace = true }
+anyhow = { workspace = true }
+log = {workspace = true}
+chrono = {workspace = true}
+bincode = {workspace = true}
+base64 = {workspace = true}
+tokio-postgres = { version = "0.7.8", features = ["with-chrono-0_4"] }
