+++ history/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod block_stores;
 pub mod history;
+pub mod postgres;
